refactor: create main function to call individual functions
A main function was created to allow for a more modular way of calling and testing functions individually. A parameter was added to the get stock required function so that the relevant worksheet could be passed to the function, considering scope of being called from another function. This commit also saw the removal of the function that was made redundant by a gspread method.